# docs/swagger-yaml-client-generated/swagger.yaml
...
swagger: "2.0"
info:
  description: API for demartment of computer engineering and cybersecurity
  version: 1.0.0
  title: CE and CS API
  termsOfService: terms
  contact:
    name: LNTU
    url: http://localhost.com
    email: admin@example.com
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: virtserver.swaggerhub.com
basePath: /lntu_team/lntu_api/1.0.0
schemes:
- https
- http
paths:
...
  /extnews/{id}/:
    get:
      summary: Returns info about news by ID.
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        minimum: 1
      responses:
        200:
          description: A news object
          schema:
            $ref: '#/definitions/News'
        404:
          description: Not found
    post:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        minimum: 1
      responses:
        405:
          description: Method not allowed
    put:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        minimum: 1
      - in: body
        name: news
        description: News updating
        required: false
        schema:
          $ref: '#/definitions/News'
      responses:
        200:
          description: OK
    delete:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        minimum: 1
      responses:
        200:
          description: OK
...
definitions:
...
  News:
    type: object
    properties:
      id:
        type: integer
        example: 1
        description: A news id
      title:
        type: string
        example: News from NASA
        description: A news title
      description:
        type: string
        example: Lorem ipsum dolor sit amet habat yorum aves
        description: A news main data
      news_link:
        type: string
        format: URL
        example: http://news.com/id=1
        description: URL to news source
      images_link:
        type: string
        format: binary
      is_checked:
        type: boolean
        example: true
        description: Is checked by moderator
...